DRTII-505 Address PR comments
