+++ client/src/main/scala/drt/client/components/styles/CommonStyles.scala
@@ -1,12 +1,15 @@
 package drt.client.components.styles
 
-import drt.client.components.styles.CssSettings._
+import drt.client.components.styles.ScalaCssImplicits.CssSettings._
 import io.kinoplan.scalajs.react.material.ui.core.colors
-import io.kinoplan.scalajs.react.material.ui.core.styles.{PaletteOptions, ThemeOptions, TypographyOptions, createMuiTheme}
+import io.kinoplan.scalajs.react.material.ui.core.styles._
+import scalacss.ScalaCssReactImplicits
 import scalacss.internal.mutable.StyleSheet
 
+class CommonStyle extends StyleSheet.Inline with ScalaCssReactImplicits {
 
-class CommonStyle extends StyleSheet.Inline {
+  import ScalaCssImplicits._
+  import dsl._
 
   def theme = createMuiTheme(
     options = ThemeOptions(
