+++ client/src/main/scala/drt/client/components/styles/ScenarioSimulationStyle.scala
@@ -1,6 +1,6 @@
 package drt.client.components.styles
 import scalacss.internal.mutable.StyleSheet
-import CssSettings._
+import ScalaCssImplicits.CssSettings._
 
 
 case class ScenarioSimulationStyle(common: CommonStyle = DefaultCommonStyle) extends StyleSheet.Inline {
