+++ client/src/test/scala/drt/client/components/charts/ChartPropsSpec.scala
@@ -75,8 +75,6 @@ object ChartPropsSpec extends TestSuite {
             |    }
             |}""".stripMargin)
 
-      println(jsonString(result))
-      println(jsonString(expected))
       assert(jsonString(result) == jsonString(expected))
     }
